requirements.txt
python-dotenv==1.0.1
selenium==4.19.0
python-dateutil==2.9.0
pyyaml==6.0.1
